permissions-policy/permissions-policy-opaque-origin-history.https.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script>

    function post_message_promise(target, message) {
      return new Promise((resolve, _) => {
        window.addEventListener('message', e => {
          resolve(e.data);
        }, { once: true });
        target.postMessage(message, '*');
      });
    }

    promise_test(() => {
      // - opaque-origin-history1.html navigates itself to opaque-origin-history2.html.
      // - opaque-origin-history2.html call window.history.back() to navigate
      // back to opaque-origin-history1.html
      // - opaque-origin-history1.html should still be able to access fullscreen
      // feature after the history.back() navigation.
      const iframe = document.createElement('iframe');
      // sandbox iframe so that it has opaque origin.
      iframe.sandbox = 'allow-scripts';
      iframe.src = 'resources/opaque-origin-history1.sub.https.html';
      iframe.allow = "fullscreen 'src'";
      document.body.appendChild(iframe);

      return new Promise((resolve, _) => iframe.onload = resolve)
        .then(() => post_message_promise(iframe.contentWindow, 'fullscreen enabled?'))
        .then(message => assert_equals(message, 'fullscreen enabled in opaque-origin-history1.html'))
        .then(() => post_message_promise(iframe.contentWindow, 'fullscreen enabled after history navigation?'))
        .then(message => assert_equals(message, 'fullscreen enabled in opaque-origin-history1.html'));
    });
  </script>
</body>
